+++ recipes/app.rb
@@ -28,21 +28,12 @@ template "#{app_dir}/.kitchen.yml" do
   helpers(ChefDK::Generator::TemplateHelper)
 end
 
-directory "#{app_dir}/test/integration/default/serverspec" do
+directory "#{app_dir}/test/integration/default" do
   recursive true
 end
 
-directory "#{app_dir}/test/integration/helpers/serverspec" do
-  recursive true
-end
-
-cookbook_file "#{app_dir}/test/integration/helpers/serverspec/spec_helper.rb" do
-  source 'serverspec_spec_helper.rb'
-  action :create_if_missing
-end
-
-template "#{app_dir}/test/integration/default/serverspec/default_spec.rb" do
-  source 'serverspec_default_spec.rb.erb'
+template "#{app_dir}/test/integration/default/default_spec.rb" do
+  source 'inspec_default_spec.rb.erb'
   helpers(ChefDK::Generator::TemplateHelper)
   action :create_if_missing
 end